+++ src/render/dom/createHTMLElementRenderer.ts
@@ -1,4 +1,5 @@
 import { HTMLElementRenderer } from "../types";
+import { setStyleProp } from "./setStyleProp";
 
 export const createHTMLElementRenderer =
   (): HTMLElementRenderer<HTMLElement> => {
@@ -31,14 +32,3 @@ export const createHTMLElementRenderer =
       },
     };
   };
-
-export const setStyleProp = (node: HTMLElement, styleValue: any) => {
-  for (const [key, value] of Object.entries(styleValue as any) as any) {
-    let v = value;
-
-    if (typeof v === "number") {
-      v = `${v}px`;
-    }
-    node.style[key] = v;
-  }
-};
